requirements.txt
pymavlink>=2.4.20
websockets>=11.0
pyyaml 